// OneDrive/Documentos/PROJETOS/PESSOAL/GESTÃO/TRG Nexus/trg-nexus---desenvolvimento (1)/src/enums.ts
export enum AppView {
    DASHBOARD = 'dashboard',
    PATIENTS = 'patients',
    THERAPY = 'therapy',
    AGENDA = 'agenda',
    FINANCIAL = 'financial',
    MARKETING = 'marketing', // New Module
    REPORTS = 'reports',
    SETTINGS = 'settings',
}

// Subscription Plans
export enum TherapistPlan {
    TRIAL = 'trial',
    INICIANTE = 'iniciante',
    PROFISSIONAL = 'profissional',
    CLINICA = 'clinica',
}